docs(components): better props, methods and events
Add a footer on each section that expands information. On props and
methods, it links to inherited properties or events. On events, it
informs the user that native events are passed through.

These notices are always displayed, regardless if the components has
props, methods or emits events on his own.

Also, fix typos.