Clean up bounds object/array test
The altered tests used to expect the mock function to be called with a
specific object, the expected value of which depends upon side effects
of a call to the unmocked underlying Leaflet method. These changes leave
the tests expecting only those things explicitly controlled by the mock
and by Vue2Leaflet.